+++ Model/Stock.go
@@ -4,6 +4,6 @@ import "time"
 
 type Stock struct {
 	SID       string    `gorm:"primary_key"`
-	SQuantity int8      `gorm:"NOT NULL"`
+	SQuantity float64   `gorm:"NOT NULL"`
 	SUpdated  time.Time `gorm:"DEFAULT:current_timestamp"`
 }
